Make it work!
* Actually call the play routine
* The chiptune engine uses commands in "packed" format. Adjust the code
to use the correct commands.
* Fix types (char may not always be signed)
* Tweak the colors a bit to improve readability
* Implement file loading in emulator for easier testing (not yet on real
hardware)
* Don't clear the whole screen on each frame (this flickers and uses
lots of CPU). We still redraw everything, and CPU load by game_frame is
still too high (you can see the cursor blink because of this). Need to
optimize redraws to only when something has changed. As we don't clear
the screen, for now add a marker line at the end of song/instr list and
be sure to clear the "*" position marker